Revert "af"
This reverts commit 2cefdb43bfb93063c446cfeb25a5d219ff681931.

+++ Test-painsfulsmile/Assets/Scripts/Player/PlayerMove.cs
@@ -6,7 +6,7 @@ public class PlayerMove : MonoBehaviour
 {
 
     Rigidbody2D rb;
-    public float speed =2;
+    public int speed =2;
     public int speedRot = 50;
 
     ChangeAssets SailScript;
